fix(settings): Expose run period and north on all settings
I realized that the north input is critical for anyone working on a real project in an office (I don't think I have ever worked on a project were project north was the same as true north). Also, now that I realized the can just treat the run period as an analysis period string, this will be useful to expose on certain recipes.