ci(deps): Enable automerging for production deps
This is going to save me a huge amount of time and is desirable as long as the merge doesn't result in a new release.

We should remove this if we ever decide that we want to use "fix" tags for updates to requirements. But I think we'll reserve releases just for the cases where we add new features.